+++ juniper-eager-loading/src/lib.rs
@@ -438,7 +438,7 @@
 //! [Juniper]: https://github.com/graphql-rust/juniper
 //! [juniper-from-schema]: https://github.com/davidpdrsn/juniper-from-schema
 
-#![doc(html_root_url = "https://docs.rs/juniper-eager-loading/0.3.0")]
+#![doc(html_root_url = "https://docs.rs/juniper-eager-loading/0.3.1")]
 #![deny(
     missing_docs,
     dead_code,